+++ R/detect.R
@@ -1,5 +1,5 @@
 # DATA CLEANING: DETECT
-#' @include AllClasses.R AllGenerics.R
+#' @include AllGenerics.R
 NULL
 
 # Count ========================================================================