+++ R/predicates.R
@@ -297,52 +297,66 @@ is_symmetric <- function(x) {
 }
 
 # Graph ========================================================================
-#' Graph Predicates
-#'
-#' `is_dag()` checks if a graph has a topological ordering (i.e. is a directed
-#' acyclic graph) using Kahn's algorithm.
-#' @param x An adjacency [`matrix`] to be tested.
-#' @return A [`logical`] scalar.
-#' @references
-#'  Kahn, A. B. (1962). Topological sorting of large networks. *Communications
-#'  of the ACM*, 5(11), p. 558-562. \doi{10.1145/368996.369025}.
-#' @family predicates
-#' @name predicate-graph
-#' @rdname predicate-graph
-#' @keywords internal
+# Graph Predicates
+#
+# `is_dag()` checks if a graph has a topological ordering (i.e. is a directed
+# acyclic graph) using Kahn's algorithm.
+# @param x An adjacency [`matrix`] to be tested.
+# @return A [`logical`] scalar.
+# @references
+#  Kahn, A. B. (1962). Topological sorting of large networks. *Communications
+#  of the ACM*, 5(11), p. 558-562. \doi{10.1145/368996.369025}.
+# @family predicates
+# @name predicate-graph
+# @rdname predicate-graph
+# @keywords internal
 NULL
 
-#' @export
-#' @rdname predicate-graph
-is_dag <- function(x) {
-  # Get edges
-  G <- matrix2edges(x)
-  # Find nodes which have no incoming edges
-  S <- which(colSums(x) == 0)
-  # List that will contain the sorted elements
-  L <- list()
+# @export
+# @rdname predicate-graph
+# is_dag <- function(x) {
+#   # Get edges
+#   G <- matrix2edges(x)
+#   # Find nodes which have no incoming edges
+#   S <- which(colSums(x) == 0)
+#   # List that will contain the sorted elements
+#   L <- list()
+#
+#   if (length(S) == 0)
+#     return(FALSE)
+#
+#   k <- 1L
+#   while (k == 1 || length(S) != 0) {
+#     # Remove a node n from S and add n to tail of L
+#     n <- S[[1]]
+#     S <- S[-1]
+#     L <- append(L, n)
+#     # For each node m with an edge e from n to m
+#     e <- which(G[, 1] == n)
+#     m <- G[e, 2]
+#     # Do remove edge e from the graph
+#     G <- G[-e, , drop = FALSE]
+#     # If m has no other incoming edges then insert m into S
+#     if (nrow(G) != 0) {
+#       m <- m[!(m %in% G[, 2])]
+#       if (length(m) != 0)
+#         S <- append(S, m)
+#     }
+#     k <- k + 1
+#   }
+#   return(nrow(G) == 0)
+# }
 
-  if (length(S) == 0)
-    return(FALSE)
-
-  k <- 1L
-  while (k == 1 || length(S) != 0) {
-    # Remove a node n from S and add n to tail of L
-    n <- S[[1]]
-    S <- S[-1]
-    L <- append(L, n)
-    # For each node m with an edge e from n to m
-    e <- which(G[, 1] == n)
-    m <- G[e, 2]
-    # Do remove edge e from the graph
-    G <- G[-e, , drop = FALSE]
-    # If m has no other incoming edges then insert m into S
-    if (nrow(G) != 0) {
-      m <- m[!(m %in% G[, 2])]
-      if (length(m) != 0)
-        S <- append(S, m)
-    }
-    k <- k + 1
-  }
-  return(nrow(G) == 0)
-}
+# matrix2edges <- function(from) {
+#   edges <- matrix(data = NA, nrow = 0, ncol = 2)
+#   nodes <- seq_len(nrow(from))
+#   for (i in nodes) {
+#     to <- which(from[i, ])
+#     if (length(to) != 0) {
+#       e <- cbind(form = i, to = to)
+#       edges <- rbind(edges, e)
+#     }
+#   }
+#
+#   edges
+# }