+++ R/replace.R
@@ -1,5 +1,5 @@
 # DATA CLEANING: REPLACE
-#' @include AllClasses.R AllGenerics.R
+#' @include AllGenerics.R
 NULL
 
 # Replace ======================================================================