+++ R/zzz.R
@@ -1,7 +1,6 @@
 .onLoad <- function(libname, pkgname) {
   op <- options()
   op.arkhe <- list(
-    arkhe.autodetect = TRUE,
     arkhe.verbose = TRUE
   )
   toset <- !(names(op.arkhe) %in% names(op))