+++ NEWS.md
@@ -1,4 +1,7 @@
 # arkhe 0.5.0.9000
+## New classes and methods
+* Export all type and scalar predicates.
+
 ## Breaking changes
 * Remove all `*Matrix` classes and methods.